Jetson TK1: add missing PCIe-related pin configuration
The Jetson TK1 spreadsheet is missing configuration for the PCIe clk_req,
rst, and wake pins. This causes the generated pinmux tables to also omit
any configuration for these pins, which in turn causes U-Boot's and the
Linux Kernel's PCIe support to fail.

Manually add configuration for these pins. The values here match the
values found in the downstream L4T kernel, and common sense based on the
usage of these pins.

Signed-off-by: Stephen Warren <swarren@nvidia.com>