Board CSV import: Put full CSV filename in data structure
Putting the whole filename in to the data structure allows us to keep
different CSV files in different directories. This will allow explicit
separation of public and private CSV files.

The per-board data is also modified to be a dictionary so that additional
fields can be stored in the future.

Signed-off-by: Stephen Warren <swarren@nvidia.com>