board-to-kernel-dt: don't generate outer state node
The important thing for board-to-kernel-dt to generate is the content of
the pinmux "state" node in DT. The name and label of the node that
contains this information is essentially irrelevant. Since this
irrelevant name and label varies between boards in existing DTs, update
board-to-kernel-dt so it doesn't generate it. This makes it obvious that
the name and label shouldn't be modified when importing new versions of
board-to-kernel-dt's output.

Signed-off-by: Stephen Warren <swarren@nvidia.com>
Reviewed-by: Simon Glass <sjg@chromium.org>